+++ src/models/data/SeoData.php
@@ -131,6 +131,15 @@ class SeoData extends BaseObject
 			unset($config['title']);
 		}
 
+		if (array_key_exists('title', $config))
+		{
+			$config['titleRaw'] = $config['title'];
+			unset($config['title']);
+		}
+
+		if (array_key_exists('titleRaw', $config) && is_string($config['titleRaw']))
+			$config['titleRaw'] = [$config['titleRaw']];
+
 		// Backwards compatibility for descriptions in SEO v3.4.x or lower
 		if (isset($config['description']))
 		{